+++ test/cases/110-const-div-zero/main.sol
@@ -2,10 +2,10 @@ package constdivzero
 
 struct Main {
 
-    public func run(args: String...): Long {
+    public func run(args: String...): Integer {
         // A constant expression that divides by zero must still fail at
         // runtime; the optimizer must not fold it away.
-        let x: Long = 1 / 0
+        let x: Integer = 1 / 0
         return x
     }
 }
