+++ test/cases/144-compile-error-protected/main.sol
@@ -4,4 +4,4 @@ struct Account {
     protected balance: Long
 }
 
-struct Main { public func run(args: String...): Long { return 0 } }
+struct Main { public func run(args: String...): Integer { return 0 } }
