+++ test/cases/145-compile-error-delegate-non-interface/main.sol
@@ -7,4 +7,4 @@ struct Wrapper {
     delegate Thing to thing
 }
 
-struct Main { public func run(args: String...): Long { return 0 } }
+struct Main { public func run(args: String...): Integer { return 0 } }
