+++ test/cases/212-legacy-static-method-rejected/main.sol
@@ -5,5 +5,5 @@ struct X {
 }
 
 struct Main {
-    public func run(args: String...): Long { return 0 }
+    public func run(args: String...): Integer { return 0 }
 }
