+++ test/cases/56-stray-brace/main.sol
@@ -1,7 +1,7 @@
 package t
 }
 struct Main {
-    public func run(args: String...): Long {
+    public func run(args: String...): Integer {
         return 0
     }
 }
